Export some fuctions to modules. Create new modules FormUtilities and symptomCardUtilities
